refactor: merge forecast display transparency variants into base components
Absorb forecast-display-with-transparency into forecast-display.tsx
and beaches-enhanced-forecast-with-transparency into beaches-enhanced-forecast.tsx.
Transparency features gated behind optional props (default false).
Delete 2 -with-transparency files.

Phase 2B of codebase redundancy cleanup.
